+++ src/briefy/leica/models/billing_info/__init__.py
@@ -1,0 +1,1 @@
+"""Billing information package."""